Colision detection.
Ok lost added. Added two functions to the function libary. It returns
the four points that mark the vertexes of a square surrounding a sprite,
or the 4 points given by a position (x and y) a width and a height.  And
yea, added colision detection. Will go through on bus tomorrow on what
the fuck it all means.