Added multi-player LevelEntityManager handling
sounds more fancy than it is. I changed LevelEntityManager to work with
a vector of players, not just one player, hopefully paving the way for
AI intergration. I added the new character framework, but tbh it might
not be nessesary. You could easily just use the Character class (which
had all the functionality of a player controlled character) and then add
some functions to LevelEntityManager. Maybe you could (like how we work
with the weapon stats) have a enum CharacterStats, which will have like
3 elements, PlayerCharacter, Team1Ai, Team2Ai. Every Character stores
its "ID" as storing one of those enums, and then the LevelEntityManager
can go through the vector of players, and do different things wither its
a player character or a ai character?